requirements.txt
gunicorn==19.9.0
whitenoise==4.1.2
Flask==2.0.2
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
pymysql
